Add a FORWARDED telnet option.
This is similar to the X-Forwarded-For request header in HTTP.
The purpose is to provide administrators running a FORWARDED
aware proxy or webclient accurate hostnames in WHO*.

A client must indicate their desire to use this option with
IAC WILL FORWARDED and receive a IAC DO FORWARDED in response,
after which they may at any time send a sequence like:
IAB SB FORWARDED <hostname> IAC SE.

Note that this functionality is ignored if the descriptor socket
is not connected from a host found in a whitelist in config.h.

The whitelist is implemented as a static constant array, which
will waste a few bytes due to how many places config.h is #included.